Ensure the chosen variation is loaded in the current document
When the experiment is loaded in an iframe the cookies are not (always?) directly visible
from the main document. Therefore any call to analytics from the main document will NOT
include the variation. To make sure the variation is loaded we explicitely ask the cxApi
for the variation. As a side effect it will load the variation into the context of the
current document, and hence all following analytics calls will contain the experiment
parameters.