Fix NPE in getVariation by checking if experiment is loaded
If a non-existing (or not-loaded) experiment is requested then we
shouldn't fail with an NPE, but instead provide some sane error. This PR
introduces checks at the necessary places to be able to deal with
non-existing experiments, and to log them.

If an experiment isn't loaded then it will default to variation `0`, log
the warning, and throw. The error can't be caught, since it
happens in a directive, but it will end up in logs &
`unhandledPromiseRejection` callbacks.